Feature/build on windows (#6)
* biting the bullet here, adding support for windows. 1. got the correct dependencies downloaded and extracted. 2. started a transition to powershell for windows scripting. 3. start adding msvc specific changes for flags and linking. This last step is still a work in progress as the project does not build yet.

* first working Windows builds. Switch to /MT, build XED for windows now, switch Disassembler to a static lib. Also fixed warnings.

+++ src/capstoneDisassembler.cpp
@@ -42,7 +42,7 @@ CapstoneDisassembler::~CapstoneDisassembler()
     cs_close(&handle);
 }
 
-void CapstoneDisassembler::Decode(const unsigned char *code, int size) {
+void CapstoneDisassembler::Decode(const unsigned char *code, size_t size) {
 
     cs_insn *insn = nullptr;
     size_t count = decodeInstruction(code, size, insn);
@@ -55,11 +55,11 @@ void CapstoneDisassembler::Decode(const unsigned char *code, int size) {
     freeInstruction(insn, count);
 }
 
-int CapstoneDisassembler::decodeInstruction(const unsigned char *code, int size, cs_insn *&insn) {
-    int count;
+size_t CapstoneDisassembler::decodeInstruction(const unsigned char *code, size_t size, cs_insn *&insn) {
+    size_t count;
     count = cs_disasm(handle, code, size, 0, 0, &insn);
     return count;
 }
-void CapstoneDisassembler::freeInstruction(cs_insn *insn, int count) {
+void CapstoneDisassembler::freeInstruction(cs_insn *insn, size_t count) {
     cs_free(insn, count);
 }
